Update transcript.go for empty whisper answer
If the stream has various seconds without spoken words, whisper refuses to give a correct answer. The answer is just empty. No timestamp or text.
In conclusion the player gets an error 500 because the database doesn't have an entry for this videopart.
I added an if-event if the answer is empty to change the answer to the beginning of the actual timestamp for this segment, which is counted up with each segment,  to +4 seconds in my case.